Move all filter sets API calls to ExplorerFilterSetsContext
