+++ src/GuppyDataExplorer/ExplorerFilterSet/index.jsx
@@ -6,20 +6,13 @@ import 'rc-tooltip/assets/bootstrap_white.css';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import SimplePopup from '../../components/SimplePopup';
 import Button from '../../gen3-ui-component/components/Button';
-import { useExplorerConfig } from '../ExplorerConfigContext';
 import { useExplorerState } from '../ExplorerStateContext';
 import { useExplorerFilterSets } from '../ExplorerFilterSetsContext';
 import {
   FilterSetActionMenu,
   FilterSetActionForm,
 } from './FilterSetActionComponents';
-import {
-  createEmptyFilterSet,
-  truncateWithEllipsis,
-  createFilterSet,
-  updateFilterSet,
-  deleteFilterSet,
-} from './utils';
+import { createEmptyFilterSet, truncateWithEllipsis } from './utils';
 import './ExplorerFilterSet.css';
 import './typedef';
 
@@ -29,11 +22,16 @@ import './typedef';
  * @param {ExplorerFilters} prop.filter
  */
 function ExplorerFilterSet({ className, filter }) {
-  const { explorerId } = useExplorerConfig();
   const { clearFilters, updateFilters } = useExplorerState();
   const [filterSet, setFilterSet] = useState(createEmptyFilterSet());
 
-  const { filterSets, refreshFilterSets } = useExplorerFilterSets();
+  const {
+    filterSets,
+    refreshFilterSets,
+    createFilterSet,
+    deleteFilterSet,
+    updateFilterSet,
+  } = useExplorerFilterSets();
   const [isError, setIsError] = useState(false);
   useEffect(() => {
     if (!isError) refreshFilterSets().catch(() => setIsError(true));
@@ -61,7 +59,7 @@ function ExplorerFilterSet({ className, filter }) {
   }
   async function handleCreate(/** @type {ExplorerFilterSet} */ created) {
     try {
-      setFilterSet(await createFilterSet(explorerId, created));
+      setFilterSet(await createFilterSet(created));
       await refreshFilterSets();
     } catch (e) {
       setIsError(true);
@@ -71,7 +69,7 @@ function ExplorerFilterSet({ className, filter }) {
   }
   async function handleUpdate(/** @type {ExplorerFilterSet} */ updated) {
     try {
-      await updateFilterSet(explorerId, updated);
+      await updateFilterSet(updated);
       setFilterSet(cloneDeep(updated));
       await refreshFilterSets();
     } catch (e) {
@@ -82,7 +80,7 @@ function ExplorerFilterSet({ className, filter }) {
   }
   async function handleDelete(/** @type {ExplorerFilterSet} */ deleted) {
     try {
-      await deleteFilterSet(explorerId, deleted);
+      await deleteFilterSet(deleted);
       setFilterSet(createEmptyFilterSet());
       await refreshFilterSets();
       clearFilters();
