+++ src/GuppyDataExplorer/ExplorerFilterSetsContext.jsx
@@ -28,9 +28,56 @@ function fetchFilterSets(explorerId) {
 }
 
 /**
+ * @param {number} explorerId
+ * @param {ExplorerFilterSet} filterSet
+ * @returns {Promise<ExplorerFilterSet>}
+ */
+export function createFilterSet(explorerId, filterSet) {
+  return fetchWithCreds({
+    path: `${FILTER_SET_URL}?explorerId=${explorerId}`,
+    method: 'POST',
+    body: JSON.stringify(filterSet),
+  }).then(({ response, data, status }) => {
+    if (status !== 200) throw response.statusText;
+    return data;
+  });
+}
+
+/**
+ * @param {number} explorerId
+ * @param {ExplorerFilterSet} filterSet
+ */
+export function deleteFilterSet(explorerId, filterSet) {
+  return fetchWithCreds({
+    path: `${FILTER_SET_URL}/${filterSet.id}?explorerId=${explorerId}`,
+    method: 'DELETE',
+  }).then(({ response, status }) => {
+    if (status !== 200) throw response.statusText;
+  });
+}
+
+/**
+ * @param {number} explorerId
+ * @param {ExplorerFilterSet} filterSet
+ */
+export function updateFilterSet(explorerId, filterSet) {
+  const { id, ...requestBody } = filterSet;
+  return fetchWithCreds({
+    path: `${FILTER_SET_URL}/${id}?explorerId=${explorerId}`,
+    method: 'PUT',
+    body: JSON.stringify(requestBody),
+  }).then(({ response, status }) => {
+    if (status !== 200) throw response.statusText;
+  });
+}
+
+/**
  * @typedef {Object} ExplorerFilterSetsContext
  * @property {ExplorerFilterSet[]} filterSets
  * @property {() => Promise<void>} refreshFilterSets
+ * @property {(filerSet: ExplorerFilterSet) => Promise<ExplorerFilterSet>} createFilterSet
+ * @property {(filerSet: ExplorerFilterSet) => Promise<void>} deleteFilterSet
+ * @property {(filerSet: ExplorerFilterSet) => Promise<void>} updateFilterSet
  */
 
 /** @type {React.Context<ExplorerFilterSetsContext>} */
@@ -47,6 +94,9 @@ export function ExplorerFilterSetsProvider({ children }) {
         filterSets,
         refreshFilterSets: () =>
           fetchFilterSets(explorerId).then(setFilterSets),
+        createFilterSet: (filterSet) => createFilterSet(explorerId, filterSet),
+        deleteFilterSet: (filterSet) => deleteFilterSet(explorerId, filterSet),
+        updateFilterSet: (filterSet) => updateFilterSet(explorerId, filterSet),
       }}
     >
       {children}
